# Fix: Optimize no-irregular-whitespace for the common case (fixes #6116)