The bulk of the expense from running `no-irregular-whitespace` comes from visiting each `Identifier` - not the work performed while visiting, just the act of visiting.

This change optimizes the rule for the common case of not having any irregular whitespace. It's really cheap to test the entire source for the irregular characters before creating the visitors, and then only create them if there are any irregular characters. A few more changes include: creating the regexs only created once, and using `test` instead of `match` where possible, among others.

Before:

```
$ for i in {1..5}; do TIMING=1 ./bin/eslint.js --no-eslintrc --rule 'no-irregular-whitespace: 1' lib; done
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |   124.392 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |   104.545 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |   111.947 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |   115.883 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |   109.632 |   100.0%
```

After:

```
$ for i in {1..5}; do TIMING=1 ./bin/eslint.js --no-eslintrc --rule 'no-irregular-whitespace: 1' lib; done
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |     3.380 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |     3.083 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |     2.813 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |     2.931 |   100.0%
Rule                    | Time (ms) | Relative
:-----------------------|----------:|--------:
no-irregular-whitespace |     2.861 |   100.0%
```