@mysticatea 2016-05-09
Cool.
Yeah, most source codes probably don't have irregular whitespaces, so this optimization has good effects.