@ilyavolodin 2016-05-09
That's some massive perf improvement! Nice job. LGTM.