@mysticatea 2016-05-13
This PR was opened while @jquerybot is down.
Could we confirm CLA?