@alberto 2016-05-13
@zertosh we changed our CLA, so you need to sign the new one at https://contribute.jquery.org/CLA/